+++ dataforge-workspace/src/jvmTest/kotlin/space/kscience/dataforge/workspace/FileDataTest.kt
@@ -23,7 +23,7 @@ import kotlin.test.assertEquals
 
 class FileDataTest {
     val dataNode = DataTree<String> {
-        branch("dir") {
+        putAll("dir") {
             wrap("a", "Some string") {
                 "content" put "Some string"
             }